requirements.txt
Django==4.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
